sv_timestamp_analysis: add hypervisor and subscriber name arguments
The name of the subscriber and the hypervisor is used in the graph and
in the adoc report.
The current method is to read the name of the sv_file and try to deduce
the name of the machine.
Is it simpler to enter the name of the machine as argument and fallback
to the sv file name if the argument is not given.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>