sv_timestamp_analysis: check for packets reordering
sv_timestamp_logger handles an iteration counter. This counter
increments when the current SV has a counter lower than the
previous one.
This mechanism allows counting the number of batches of SV that were sent.

However, if sample values are missorderd, the counter will increment
even when only one batch of sample values is sent.
In that case, it is impossible to link the received SV to the published
one.

This commit introduces a check to verify that two SV files have the same
number of streams and iterations. If not, it is mathematically
impossible to link SV correctly to compute latencies. (or at least I
didn't find a way to do it)

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>