sv_timestamp_analysis: rewrite sv drop handling function
Latency computation is straight-forward when there is no sv drop.

In case of sv drop, the two tables (publisher and subscriber) doesn't
contain the same number of data.
This commit uses the pandas dataframe merge function with the "inner"
strategy to merge the two tables. This strategy only keeps the data
presents in the two tables, thus droping the lines where the SV was
dropped.

Converting datas from numpy to pandas (or revert) takes time, so :
- pandas is only used in case of sv_drop
- in that case, pandas also compute the latency

May be pandas should be used in the entire script instead of numpy, I
don't have time to test that now.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>