sv_timestamp_analysis: display ID of the recorded SV stream
Previously, the generated report was displaying the number of recorder
SV streams next to the min, max and average statistics.
However, as for the displayed graph, these statistics are only related
to the first SV stream recorded.

Replace the number of recorded SV streams by the ID of the SV stream
used for computing the statistics. That way it is specified which stream
in actually related to the displayed statistics.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>