sv_timestamp_analysis: iterate over values in save_latency_histogram()
In save_latency_histogram(), instead of iterating with a length range of
"streams", iterate on values and streams as they have the same size.
Each element of "values" is an array of latencies which are associated
with a string index specified via the command line arguments.

This will ease reusing theses values in future commits.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>