sv_timestamp_analysis: set histogram filename to the right stream index
Now that save_latency_histogram() iterates over the stream indexes given
via command line, the name of the generated histogram files might not
match the filename in the .adoc template strings.

Histogram filenames contain the stream index as given in the command
line, and the template looks for the histogram of stream index 0.
However, if the first stream index is not 0, e.g. when passing
"--stream 3..5", there won't be any histogram for the stream index 0.

Modify the values given to the template to use the first given stream
index.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>